Add content_hash tracking and extract composer package (Phase 1)
Foundation for incremental sync: compute and store content hashes at
update time so the future sync step can diff against deployed_hash
to find what needs uploading to R2.

- Migration 024: add content_hash, deployed_hash, content_changed_at
  columns to packages table
- Extract internal/composer/ from internal/repository/ (format.go,
  serialize.go, hash.go) — pure functions for Composer serialization
  and hashing with no I/O
- Add SerializePackage() for tagged/dev file splitting and
  HashVersions() for lightweight content hashing
- Compute content_hash in update step after NormalizeAndStoreVersions;
  only advance content_changed_at when hash actually changes
- Hash includes trunk_revision so dev-trunk SVN reference changes
  trigger re-sync
- repository/ re-exports from composer/ for builder.go compatibility
