+++ internal/packages/package_test.go
@@ -40,6 +40,9 @@ func setupTestDB(t *testing.T) *sql.DB {
 			last_synced_at TEXT,
 			last_sync_run_id INTEGER,
 			trunk_revision INTEGER,
+			content_hash TEXT,
+			deployed_hash TEXT,
+			content_changed_at TEXT,
 			wp_packages_installs_total INTEGER NOT NULL DEFAULT 0,
 			wp_packages_installs_30d INTEGER NOT NULL DEFAULT 0,
 			last_installed_at TEXT,
